+++ src/main/java/com/nhnacademy/illuwa/d_book/tag/Tag.java
@@ -1,4 +1,4 @@
-package com.nhnacademy.illuwa.d_book.entity.tag;
+package com.nhnacademy.illuwa.d_book.tag;
 
 import jakarta.persistence.*;
 import lombok.Getter;
